+++ cpp/order/RotationalAutocorrelation.cc
@@ -48,11 +48,11 @@ inline std::complex<float> RotationalAutocorrelation::hypersphere_harmonic(const
     unsigned int bound = std::min(a, b);
     for (unsigned int k = (a+b < l ? 0 : a+b-l); k <= bound; k++)
     {
-        float fact_product = m_factorials.get()[k] * m_factorials.get()[l + k - a - b] * m_factorials.get()[a - k] * m_factorials.get()[b - k];
+        float fact_product = m_factorials[k] * m_factorials[l + k - a - b] * m_factorials[a - k] * m_factorials[b - k];
         sum_tracker += cpow(xi_conj, k) * cpow(zeta, b - k) * cpow(zeta_conj, a - k) * cpow(-xi, l + k - a - b) / fact_product;
     }
     sum_tracker
-        *= std::sqrt(m_factorials.get()[a] * m_factorials.get()[l - a] * m_factorials.get()[b] * m_factorials.get()[l - b] / (float(l) + 1));
+        *= std::sqrt(m_factorials[a] * m_factorials[l - a] * m_factorials[b] * m_factorials[l - b] / (float(l) + 1));
     return sum_tracker;
 }
 
@@ -63,13 +63,8 @@ void RotationalAutocorrelation::compute(const quat<float>* ref_ors, const quat<f
     assert(N > 0);
     assert(ref_ors.size == ors.size);
 
-    // Resize array if needed. No need to reset memory, we can do that in the loop (in parallel).
-    if (N != m_N)
-    {
-        m_RA_array = std::shared_ptr<std::complex<float>>(new std::complex<float>[N],
-                                                          std::default_delete<std::complex<float>[]>());
-        m_N = N;
-    }
+    m_N = N;
+    m_RA_array.prepare(m_N);
 
     // Precompute the hyperspherical harmonics for the unit quaternion. The
     // default quaternion constructor gives a unit quaternion. We will assume
@@ -97,7 +92,7 @@ void RotationalAutocorrelation::compute(const quat<float>* ref_ors, const quat<f
             std::complex<float> zeta = std::complex<float>(qq_1.v.z, qq_1.s);
 
             // Loop through the valid quantum numbers.
-            m_RA_array.get()[i] = std::complex<float>(0, 0);
+            m_RA_array[i] = std::complex<float>(0, 0);
             unsigned int uh_index = 0;
             for (unsigned int a = 0; a <= m_l; a++)
             {
@@ -105,7 +100,7 @@ void RotationalAutocorrelation::compute(const quat<float>* ref_ors, const quat<f
                 {
                     std::complex<float> combined_value = unit_harmonics[uh_index]
                         * hypersphere_harmonic(xi, zeta, m_l, a, b);
-                    m_RA_array.get()[i] += combined_value;
+                    m_RA_array[i] += combined_value;
                     uh_index += 1;
                 }
             }
@@ -115,7 +110,7 @@ void RotationalAutocorrelation::compute(const quat<float>* ref_ors, const quat<f
     float RA_sum(0);
     for (unsigned int i = 0; i < N; i++)
     {
-        RA_sum += std::real(m_RA_array.get()[i]);
+        RA_sum += std::real(m_RA_array[i]);
     }
     m_Ft = RA_sum / N;
 };